% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_compact_fit}
\alias{export_compact_fit}
\title{Export model fit to compact format}
\usage{
export_compact_fit(fit)
}
\arguments{
\item{fit}{a `disclapmixfit` object}
}
\description{
Includes only the most important aspects (parameters). 
For example useful if exporting fits for prediction to JSON.
}
\examples{
data(danes)
db <- as.matrix(danes[rep(1:nrow(danes), danes$n), 1:(ncol(danes)-1)])
fit <- disclapmix(db, clusters = 4L)
str(export_compact_fit(fit), 1)
}
